Rework imputation-server-filter workflow to use bcftools only
Because the info files can be read by bcftools, it's more straightforward
to do all of the filtering and subsetting using bcftools. This means
changing the workflow input such that the user specifes an array of
arrays. Remove the Dockerfile, since a custom docker image was no
longer needed. Update the test data and checking Rmd. Update docs.